Use Lodash as well as Underscore
the "_" var still equals Underscore so all the original code still works. You can use Lodash with "lodash". You could migrate the whole thing to Lodash if you like - I thought about doing this myself, but if you're not familiar with Lodash you might not be happy I did this. I would recommend switching. If you're not familiar with Lodash the doing a migration from Underscore is a good way to learn the differences.