Fix bugs related to remove actions
Prior to fix if there's a remove action that's { instant: false } in a txn with at some { instant: true } actions, the removed doc would not be saved. If the remove action was in txn with other { instant: false } items only - it would work because on line 1444 there wouldn't be an existing txn yet, so the items - including the removed doc added on line 1438 - would be saved at that point. This is fixed now saving items on line 1475. 

Also - the block beginning on line 1432 is now executed only if the remove is  { instant: false }. Prior to fix this would be setting doc=undefined on line 1438.