Don't set the "deleted" field of a saved hard-deleted document.
This field caused the document to be restored as soft-deleted, which is
bogus even if the only concrete consequence for an app that doesn't look
at the "deleted" field is that a subsequent remove of the document will
fail with "Document not found for removal".  It's unclear to me what
purpose setting the field ever served.